requirements.txt
django>=4.0
gunicorn
psycopg2-binary
python_dotenv
dj_database_url
django-tinymce
whitenoise
djangorestframework

